[#148359271] Apply to_json if data type is Hash
By default, body (params) is of text format thus is sent over api request as string.
Recent changes to ETL converts the database type of job.body from text to json.
As a result, params is now sent over api request as a hash/json and is breaking
Inventory syncs.